Bug fix
Surprising that this error had not previously surfaced. I think this is because the erroneous output of this function was monotonically related to the true value for the circumstance of inverse ray tracing a point from the iris stop. The error was revealed when I started to find rays for the glint path.

Signed-off-by: gkaguirre <gkaguirre@me.com>