Handle orientation case
Encountered circumstance in which the axis orientations derived from the Euler angles was not in the defined set of orientations. Now have all permutations covered. I am not certain that this function offers a solution that is correct across all combinations of quadric surfaces and rotations, but it does seem to handle the cases we encounter in the model eye.

Signed-off-by: gkaguirre <gkaguirre@me.com>