calculate fixation pose
I wrote this to understand a bit more about ocular parallax. Because the center of rotation of the eye is different from the optical center (“nodal point”) of the eye, the visual angle of a world object and the angular rotation of the eye required to bring that object into fixation are different. It’s a small effect, but I wanted to get a handle on it.

Signed-off-by: gkaguirre <gkaguirre@me.com>