Major update to coordinate system
The world coordinate system is now defined with respect to the eye in “primary position”. Previously, the coordinate system was defined relative to the alignment of the longitudinal axis of the eye with the optical axis of an eye tracking camera. To support the updated coordinate approach, the rotation of the eye tracking camera is now supported in sceneGeometry. If not specified, the camera is rotated so that it is aimed at the rotation center of the eye.
Substantial comment and code polish throughout. Header comments now make clear that the z axis of the eye and world coordinate systems are defined by the longitudinal axis of the eye, which itself is one of the semi-axes of the ellipsoidal vitreous chamber.